Se mejoro la interfaz de la aplicacion y ahora se maneja mejor los errores al momento de ejecutar el script.

+++ src/utils/organize_file.py
@@ -12,19 +12,27 @@ def get_folded_path():
         return None
 
 def organize_files(initial_path, final_path):
-    for filename in os.listdir(initial_path):
-        name, type = os.path.splitext(filename)
+      try:
+           for filename in os.listdir(initial_path):
+            source = os.path.join(initial_path, filename)
+            name, type = os.path.splitext(filename)
 
-        if not os.path.exists(f'{final_path}\{type}'):
-            new_folder = type.replace('.', '').upper()
-            Path(f'{final_path}\{new_folder}').mkdir(exist_ok=True)
-            shutil.move( 
-                f"{initial_path}\{filename}",
-                f"{final_path}\{new_folder}"
-            )
+            if os.path.isdir(source):
+                continue
 
-        else:
-            shutil.move(
-                f"{initial_path}\{filename}",
-                f"{final_path}\{new_folder}"
-            )
+            if not os.path.exists(f'{final_path}\{type}') and type != '.ini':
+                new_folder = type.replace('.', '').upper()
+                Path(f'{final_path}\{new_folder}').mkdir(exist_ok=True)
+                shutil.move( 
+                    f"{initial_path}\{filename}",
+                    f"{final_path}\{new_folder}"
+                )
+
+            else:
+                shutil.move(
+                    f"{initial_path}\{filename}",
+                    f"{final_path}\{new_folder}"
+                )
+      
+      except TypeError as e: 
+          return {"success": False, "error": "No se encontró la carpeta origen"}
